+++ cookbooks/esri-tomcat/metadata.rb
@@ -4,14 +4,14 @@ maintainer_email 'contracts@esri.com'
 license          'Apache 2.0'
 description      'Installs/Configures esri-tomcat'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.2.1'
+version          '0.2.2'
 chef_version     '>= 14.0'
 
 depends          'tomcat', '>= 3.2.0'
 depends          'openssl', '~> 8.5'
 depends          'iptables', '~> 7.1'
 
-%w(ubuntu redhat centos oracle suse).each do |os|
+%w(ubuntu redhat centos oracle suse rocky almalinux).each do |os|
   supports os
 end