+++ cookbooks/arcgis-enterprise/CHANGELOG.md
@@ -3,6 +3,11 @@ arcgis-enterprise cookbook CHANGELOG
 
 This file is used to list changes made in each version of the arcgis-enterprise cookbook.
 
+4.2.0
+-----
+- Added support for ArcGIS Enterprise 11.2
+- Added support for Rocky Linux and AlmaLinux platforms
+
 4.1.0
 -----
 - Added support for ArcGIS Enterprise 11.1